updating tests

+++ tests.cpp
@@ -7,5 +7,5 @@
 TEST_CASE("Length") {
     Coord3D p = {10, 20, 30};
     std::cout << length(&p) << "\n";
-    CHECK(length(&p)==37.4166);
+    CHECK(double(length(&p))==double(37.4166));
 }
